update latest version

+++ src/Microweber/Utils/Backup/DatabaseWriter.php
@@ -15,6 +15,7 @@ use Microweber\Utils\Backup\Traits\DatabaseMediaWriter;
 use Microweber\Utils\Backup\Loggers\BackupImportLogger;
 use Microweber\Utils\Backup\Traits\DatabaseModuleWriter;
 use Microweber\Utils\Backup\Traits\DatabaseTaggingTaggedWriter;
+use QueryPath\Exception;
 
 /**
  * Microweber - Backup Module Database Writer
@@ -118,7 +119,7 @@ class DatabaseWriter
 	
 	private function _saveItemDatabase($item) {
 		
-		if ($this->overwriteById  && isset($item['id'])) {
+		if ($this->overwriteById && isset($item['id'])) {
 			
 			// We will overwrite content by id from our db structure
 			$dbSelectParams = array();
@@ -251,9 +252,12 @@ class DatabaseWriter
 	 * @param array $item
 	 */
 	private function _saveItem($item) {
-		
+
 		$savedItem = $this->_saveItemDatabase($item);
-		
+        if ($this->overwriteById) {
+            return true; 
+        }
+
 		if ($savedItem) {
 			$this->_fixRelations($savedItem);
 			$this->_fixParentRelationship($savedItem);
@@ -288,7 +292,16 @@ class DatabaseWriter
 		var_dump($items);
 		return; */
 
+        if (isset($this->content->__table_structures)) {
+            app()->database_manager->build_tables($this->content->__table_structures);
+        }
+
 		foreach ($this->content as $table=>$items) {
+
+            if (!\Schema::hasTable($table)) {
+                continue;
+            }
+
 			if (!empty($items)) {
 				foreach($items as $item) {
 					$item['save_to_table'] = $table;
@@ -315,6 +328,10 @@ class DatabaseWriter
 			$this->_finishUp('runWriterWithBatchNothingToImport');
 			return array("success"=>"Nothing to import.");
 		}
+
+		if (isset($this->content->__table_structures)) {
+		    app()->database_manager->build_tables($this->content->__table_structures);
+        }
 		
 		//$importTables = array('users', 'categories', 'modules', 'comments', 'content', 'media', 'options', 'calendar', 'cart_orders');
 		//$importTables = array('content', 'categories');
@@ -323,7 +340,11 @@ class DatabaseWriter
 		// All db tables
 		$itemsForSave = array();
 		foreach ($this->content as $table=>$items) {
-			
+
+            if (!\Schema::hasTable($table)) {
+                continue;
+            }
+
 			if (in_array($table, $excludeTables)) {
 				continue;
 			}
@@ -403,7 +424,10 @@ class DatabaseWriter
                 if ($table == 'users' || $table == 'users_oauth' || $table == 'system_licenses') {
                     continue;
                 }
-                \DB::table($table)->truncate();
+                if (\Schema::hasTable($table)) {
+                    BackupImportLogger::setLogInfo('Truncate table: ' . $table);
+                    \DB::table($table)->truncate();
+                }
             }
         }
     }
